requirements.txt
# Store Item Detection - Python Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
albumentations>=1.3.0
pyyaml>=6.0

# Deep learning utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Object detection frameworks (optional - uncomment as needed)
# ultralytics>=8.0.0  # For YOLO models
# timm>=0.9.0  # For various backbones
# mmdet>=3.0.0  # For MMDetection framework

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Export and deployment
onnx>=1.14.0
onnxruntime>=1.15.0

# Utilities
scikit-learn>=1.3.0
pandas>=2.0.0
